Forgot to convert some token representations from int to byte. Shouldn't provoke build errors in mono environments anymore.

+++ DParser2/Dom/Nodes/DNode.cs
@@ -69,7 +69,7 @@ namespace D_Parser.Dom
 			return false;
 		}
 
-        public bool ContainsAttribute(params int[] Token)
+        public bool ContainsAttribute(params byte[] Token)
         {
             return Modifier.ContainsAttribute(Attributes, Token);
         }
